24.01.09_회원가입완료(일반회원, 사업자회원)

+++ nosolorice/src/main/java/com/nosolorice/app/domain/businessUser/BusinessUser.java
@@ -9,7 +9,7 @@ public class BusinessUser {
 	private String pass;						// 비밀번호
 	private String name;					// 사장님 이름
 	private String phone;					// 사업장 가게번호
-	private int businessNumber;			// 사업자등록번호
+	private long businessNumber;			// 사업자등록번호
 	private String businessName;		// 상호명
 	private String businessPicture;		// 사업자등록증 사본파일
 	private String email;					// 이메일
@@ -21,7 +21,7 @@ public class BusinessUser {
 	private String ypoint;					// 경도
 	private String businessProfile;		// 프로필사진파일
 	private String bankName;				// 계좌은행
-	private int bankNumber;				// 계좌번호
+	private long bankNumber;				// 계좌번호
 	private int deposit;					// 가게 예약금
 	private String openTime;				// 오픈시간
 	private String closeTime;				// 마감시간
@@ -107,11 +107,11 @@ public class BusinessUser {
 		this.phone = phone;
 	}
 
-	public int getBusinessNumber() {
+	public long getBusinessNumber() {
 		return businessNumber;
 	}
 
-	public void setBusinessNumber(int businessNumber) {
+	public void setBusinessNumber(long businessNumber) {
 		this.businessNumber = businessNumber;
 	}
 
@@ -203,11 +203,11 @@ public class BusinessUser {
 		this.bankName = bankName;
 	}
 
-	public int getBankNumber() {
+	public long getBankNumber() {
 		return bankNumber;
 	}
 
-	public void setBankNumber(int bankNumber) {
+	public void setBankNumber(long bankNumber) {
 		this.bankNumber = bankNumber;
 	}
